Don't notify if buffer is active
This patch adds a new configuration option, `inactive only`, that
allows a user to disable notifications for a buffer that is active.

For example, if a user has a private message buffer active, they
may not want pushbullet messages sent to them from conversations
in that buffer.

Users can set `inactive_only` to `0` if they still want pushbullet
notifications even in their active buffers.